Create a seperate, reduced, view within view_total for when the user has added participants.
This removes the IPVs visibility and instead just reports on total unallocated portion and total tip and overal total. THe portion total is correctly transferred between the separate views when a swap is made. Not all cases for a swap are supported, and will need to be added when corresponding functionality (such as when the user removes participants) is added.
Still need to add functionality for updating the report fields in view_total when the total portion is added by the user.