Persist information in total_view.
Added entity and dao for storing BasicInfo, this table is only intended to store single-instance datapoints, and will be treated as such by the Database class.

Add code to MainAcitivty to utilize this data table. Callbacks to persist changes to the table, as well as value setting to restore state from the table.