Determine which mode total_view should be in on launch
Create method MainActivity::handleTotalViewVisibility that will determine and set visibility between reduced and ipv mode. Insert a call after participantAdapter is instantiated so the correct view mode will be set when the user launches the app.
This method will also be useful for when participant removal functionality is implemented to revert back to ipv mode.